@@ -4,35 +4,46 @@ declare(strict_types=1);
 
 namespace App\Service\Synchronization;
 
+use App\Entity\Greeting;
 use App\Service\Greeting\GreetingService;
+use DateTime;
+use Doctrine\ORM\EntityManagerInterface;
 
-/**
- * @author  Wolfgang Hinzmann <wolfgang.hinzmann@doccheck.com>
- * @license 2023 DocCheck Community GmbH
- */
 class GreetingSynchronizationService
 {
-    public function __construct(private DigitalFrameApiGateway $apiGateway, private GreetingService $greetingService) { }
+    public function __construct(
+        private readonly DigitalFrameApiGateway $apiGateway,
+        private readonly GreetingService $greetingService,
+        private readonly EntityManagerInterface $entityManager
+    ) {
+    }
 
     public function checkForNewGreetings(): bool
     {
-//        return true;
         return $this->apiGateway->checkGreetings();
     }
 
-    /**
-     * @param int[] $ids
-     */
-    public function markAsDisplayed(array $ids): void
+    public function synchronizeDisplayedGreetingsToServer(): void
     {
+        $greetingsToSync = $this->greetingService->getDisplayedGreetingsNeedingSync();
+        $ids = array_map(fn (Greeting $greeting) => $greeting->getRemoteId(), $greetingsToSync);
+
         $this->apiGateway->markGreetingsAsDisplayed($ids);
+
+        foreach ($greetingsToSync as $greeting) {
+            $greeting->setLastSynced(new DateTime());
+            $this->entityManager->persist($greeting);
+        }
+
+        $this->entityManager->flush();
     }
 
-    public function synchronizeGreetings(): void
+    public function synchronizeGreetingsFromServer(): void
     {
         $greetings = $this->apiGateway->getGreetings();
+
         $ids = [];
-        foreach ($greetings as $greetingData){
+        foreach ($greetings as $greetingData) {
             $this->greetingService->createByData($greetingData);
             $ids[] = $greetingData->getRemoteId();
         }
