@@ -1,6 +1,6 @@
 <?php
 
-namespace App\Service\Image\Unsplash;
+namespace App\Service\Unsplash;
 
 use App\Entity\UnsplashImage;
 use DateTime;
@@ -13,81 +13,51 @@ class UnsplashImageData
     private string $color;
     private ?DateTime $viewed;
 
-    /**
-     * @return string
-     */
     public function getName(): string
     {
         return $this->name;
     }
 
-    /**
-     * @param string $name
-     */
     public function setName(string $name): void
     {
         $this->name = $name;
     }
 
-    /**
-     * @return string
-     */
     public function getUrl(): string
     {
         return $this->url;
     }
 
-    /**
-     * @param string $url
-     */
     public function setUrl(string $url): void
     {
         $this->url = $url;
     }
 
-    /**
-     * @return string
-     */
     public function getTag(): string
     {
         return $this->tag;
     }
 
-    /**
-     * @param string $tag
-     */
     public function setTag(string $tag): void
     {
         $this->tag = $tag;
     }
 
-    /**
-     * @return string
-     */
     public function getColor(): string
     {
         return $this->color;
     }
 
-    /**
-     * @param string $color
-     */
     public function setColor(string $color): void
     {
         $this->color = $color;
     }
 
-    /**
-     * @return DateTime|null
-     */
     public function getViewed(): ?DateTime
     {
         return $this->viewed;
     }
 
-    /**
-     * @param DateTime|null $viewed
-     */
     public function setViewed(?DateTime $viewed): void
     {
         $this->viewed = $viewed;
@@ -100,6 +70,7 @@ class UnsplashImageData
         $this->setTag($image->getTag());
         $this->setUrl($image->getUrl());
         $this->setViewed($image->getViewed());
+
         return $this;
     }
 }
